requirements.txt
# Core dependencies
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Database
pymongo==4.6.1
motor==3.3.2

# AI/ML
openai==1.9.0
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0

# Media Processing
moviepy==1.0.3
pytube==15.0.0
Pillow==10.2.0

# Communication
python-telegram-bot==20.7
twilio==8.11.0
aiosmtplib==2.0.2

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
schedule==1.2.1

pytube==15.0.0
moviepy==1.0.3
openai-whisper==20231117
instabot==0.117.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0
numpy==1.24.3